restructured trader
split trader_cli from trader.py, form a separate file for cli
removed dependency from trader_cli to trader utiliti functions and trading related functions, made it a pure UI file
renamed tui.py as trader_tui.py
